move: 테스트 별 설정을 위한 디렉토리 이동 #1 - RestaurantMock scheduler/mock -> service/mock - RestaurantServiceTestConfig config -> service/config
