+++ src/test/java/com/wanted/lunchmapservice/restaurant/service/config/RestaurantServiceTestConfig.java
@@ -1,9 +1,9 @@
-package com.wanted.lunchmapservice.restaurant.config;
+package com.wanted.lunchmapservice.restaurant.service.config;
 
 import com.wanted.lunchmapservice.restaurant.mapper.RestaurantMapper;
 import com.wanted.lunchmapservice.restaurant.repository.RestaurantRepository;
-import com.wanted.lunchmapservice.restaurant.scheduler.mock.RestaurantMock;
 import com.wanted.lunchmapservice.restaurant.service.RestaurantGetService;
+import com.wanted.lunchmapservice.restaurant.service.mock.RestaurantMock;
 import org.mockito.Mockito;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
